# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TREVV API",
    "version": "2.0.0",
    "description": "Versioned, permission-scoped contracts shared by the TREVV Web, Mobile, and Desktop clients. Health, this OpenAPI document, and provider-managed /api/auth/* endpoints are public. In live mode every product request derives its user, selected organization, role, and managed scopes from a Better Auth session and active PostgreSQL memberships."
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local API"
    }
  ],
  "security": [
    {
      "SessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/v1/onboarding": {
      "get": {
        "tags": [
          "Identity"
        ],
        "operationId": "getOnboarding",
        "responses": {
          "200": {
            "description": "Recoverable onboarding progress",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingState"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/RepositoryUnavailable"
          }
        }
      },
      "put": {
        "tags": [
          "Identity"
        ],
        "operationId": "saveOnboarding",
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingDraft"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Durably saved onboarding progress",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingState"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "503": {
            "$ref": "#/components/responses/RepositoryUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "trevv.session_token",
        "description": "Better Auth session cookie. Production may apply the standard __Secure- cookie prefix; clients must obtain the cookie through the public /api/auth/* endpoints."
      }
    },
    "parameters": {
...
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Strong numeric ETag returned by the resource, from \"0\" through \"2147483647\"; for example \"3\".",
        "schema": {
          "type": "string",
          "pattern": "^\"[0-9]+\"$"
        }
      }
    },
    "headers": {
      "ETag": {
        "description": "Strong quoted PostgreSQL integer resource version, from \"0\" through \"2147483647\".",
        "schema": {
          "type": "string",
          "pattern": "^\"[0-9]+\"$"
        }
      },
...
    },
    "schemas": {
...
      "OnboardingDraft": {
        "type": "object",
        "required": [
          "step"
        ],
        "additionalProperties": false,
        "properties": {
          "step": {
            "type": "integer",
            "enum": [
              1,
              2,
              3,
              4,
              5
            ]
          },
          "organizationName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "organizationSlug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "workspaceName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "workspaceSlug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "workspaceType": {
            "type": "string",
            "enum": [
              "business",
              "brand",
              "client",
              "product",
              "department",
              "venture",
              "initiative",
              "investment",
              "campaign",
              "program",
              "project",
              "shared_function",
              "client_program",
              "journey",
              "other"
            ]
          },
          "workspaceColor": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "blueprintKey": {
            "type": "string",
            "enum": [
              "operating_business",
              "client_delivery",
              "product_initiative",
              "launch_campaign",
              "blank"
            ]
          }
        }
      },
...
      "OnboardingState": {
        "type": "object",
        "required": [
          "status",
          "step",
          "draft",
          "version",
          "updatedAt"
        ],
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "not_started",
              "in_progress",
              "completed"
            ]
          },
          "step": {
            "type": "integer",
            "enum": [
              1,
              2,
              3,
              4,
              5
            ]
          },
          "draft": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "organizationName": {
                "type": "string",
                "minLength": 2,
                "maxLength": 160
              },
              "organizationSlug": {
                "type": "string",
                "minLength": 2,
                "maxLength": 80,
                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
              },
              "workspaceName": {
                "type": "string",
                "minLength": 2,
                "maxLength": 160
              },
              "workspaceSlug": {
                "type": "string",
                "minLength": 2,
                "maxLength": 80,
                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
              },
              "workspaceType": {
                "type": "string",
                "enum": [
                  "business",
                  "brand",
                  "client",
                  "product",
                  "department",
                  "venture",
                  "initiative",
                  "investment",
                  "campaign",
                  "program",
                  "project",
                  "shared_function",
                  "client_program",
                  "journey",
                  "other"
                ]
              },
              "workspaceColor": {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              "blueprintKey": {
                "type": "string",
                "enum": [
                  "operating_business",
                  "client_delivery",
                  "product_initiative",
                  "launch_campaign",
                  "blank"
                ]
              }
            }
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "organizationId": {
            "type": "string",
            "minLength": 3,
            "maxLength": 128
          },
          "portfolioId": {
            "type": "string",
            "minLength": 3,
            "maxLength": 128
          },
          "workspaceId": {
            "type": "string",
            "minLength": 3,
            "maxLength": 128
          },
          "boardId": {
            "type": "string",
            "minLength": 3,
            "maxLength": 128
          },
          "blueprintInstanceId": {
            "type": "string",
            "minLength": 3,
            "maxLength": 128
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "properties": {
                  "currentVersion": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 2147483647
                  }
                },
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "responses": {
      "Unauthenticated": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "The resource version is stale, or an idempotency key was reused for a different request.",
        "headers": {
          "ETag": {
            "$ref": "#/components/headers/ETag"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Validation": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "A current quoted ETag is required in If-Match.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RepositoryUnavailable": {
        "description": "The live PostgreSQL repository is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}